last bits updated
