+++ src/index.ts
@@ -6,5 +6,6 @@ import * as Inspection from "./inspection";
 export * from "./common";
 export * from "./jobs";
 export * from "./lexer";
+export * from "./localization";
 export * from "./parser";
 export { Inspection };
