+++ src/localization/index.ts
@@ -1,6 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import * as Error from "./error";
-
-export { Error };
+export { Localization } from "./localization";
